+++ src/common/db/Database.py
@@ -341,7 +341,14 @@ class Database:
                     session.remove()
 
     def is_valid_setting(
-        self, setting: str, *, value: Optional[str] = None, multisite: bool = False, session: Optional[scoped_session] = None
+        self,
+        setting: str,
+        *,
+        value: Optional[str] = None,
+        multisite: bool = False,
+        session: Optional[scoped_session] = None,
+        accept_prefixed: bool = True,
+        extra_services: Optional[List[str]] = None,
     ) -> Tuple[bool, str]:
         """Check if the setting exists in the database, if it's valid and if the value is valid"""
 
@@ -354,15 +361,21 @@ class Database:
 
                 db_setting = session.query(Settings).filter_by(id=setting).first()
 
-                if not db_setting:
-                    for service in session.query(Services).with_entities(Services.id):
-                        if setting.startswith(f"{service.id}_"):
-                            db_setting = session.query(Settings).filter_by(id=setting.replace(f"{service.id}_", "")).first()
-                            multisite = True
+                if accept_prefixed and not db_setting:
+                    for service in extra_services or []:
+                        if setting.startswith(f"{service}_"):
+                            db_setting = session.query(Settings).filter_by(id=setting.replace(f"{service}_", "")).first()
                             break
 
                     if not db_setting:
-                        return False, "missing"
+                        for service in session.query(Services).with_entities(Services.id):
+                            if setting.startswith(f"{service.id}_"):
+                                db_setting = session.query(Settings).filter_by(id=setting.replace(f"{service.id}_", "")).first()
+                                multisite = True
+                                break
+
+                if not db_setting:
+                    return False, "missing"
 
                 if multisite and db_setting.context != "multisite":
                     return False, "not multisite"
@@ -421,7 +434,7 @@ class Database:
                 metadata = session.query(Metadata).with_entities(Metadata.version).filter_by(id=1).first()
                 if metadata:
                     return metadata.version
-                return "1.6.0-rc2"
+                return "1.6.0-rc3"
             except BaseException as e:
                 return f"Error: {e}"
 
@@ -451,7 +464,7 @@ class Database:
             "last_pro_plugins_change": None,
             "last_instances_change": None,
             "integration": "unknown",
-            "version": "1.6.0-rc2",
+            "version": "1.6.0-rc3",
             "database_version": "Unknown",  # ? Extracted from the database
             "default": True,  # ? Extra field to know if the returned data is the default one
         }
@@ -1447,7 +1460,7 @@ class Database:
                                     {"model": Services, "filter": {"id": server_name}, "values": {"last_update": datetime.now().astimezone()}}
                                 )
                             elif (
-                                method == service_setting["method"] or (service_setting["method"] not in ("scheduler", "autoconf") and method == "autoconf")
+                                method in (service_setting["method"], "autoconf") or (method == "scheduler" and service_setting["method"] != "autoconf")
                             ) and service_setting["value"] != value:
                                 local_changed_plugins.add(setting["plugin_id"])
 
@@ -1516,7 +1529,7 @@ class Database:
                                 local_changed_plugins.add(setting["plugin_id"])
                                 local_to_put.append(Global_values(setting_id=key, value=value, suffix=suffix, method=method))
                             elif (
-                                method == global_value.method or (global_value.method not in ("scheduler", "autoconf") and method == "autoconf")
+                                method in (global_value.method, "autoconf") or (method == "scheduler" and global_value.method != "autoconf")
                             ) and global_value.value != value:
                                 local_changed_plugins.add(setting["plugin_id"])
 
@@ -1618,7 +1631,7 @@ class Database:
                             changed_plugins.add(setting.plugin_id)
                             to_put.append(Global_values(setting_id=key, value=value, suffix=suffix, method=method))
                         elif (
-                            method == global_value.method or (global_value.method not in ("scheduler", "autoconf") and method == "autoconf")
+                            method in (global_value.method, "autoconf") or (method == "scheduler" and global_value.method != "autoconf")
                         ) and global_value.value != value:
                             changed_plugins.add(setting.plugin_id)
 
@@ -1744,7 +1757,7 @@ class Database:
                 if not custom_conf:
                     to_put.append(Custom_configs(**custom_config))
                 elif custom_config["checksum"] != custom_conf.checksum and (
-                    method == custom_conf.method or (custom_conf.method not in ("scheduler", "autoconf") and method == "autoconf")
+                    method in (custom_conf.method, "autoconf") or (method == "scheduler" and custom_conf.method != "autoconf")
                 ):
                     custom_conf.data = custom_config["data"]
                     custom_conf.checksum = custom_config["checksum"]